Modify TransitionTable to RootTable and LeafTable
This commit will split TransitionTable into RootTable, which points to
the root node of a given word's tree, and LeafTable, which contains an
array of pointers to leaf nodes of a given word.  This facilitates forward
and backward traversal.  A data member, *parent*, is also added to
TransitionNode to allow backward traversal.